+++ primitives/Cargo.toml
@@ -13,4 +13,6 @@ clap = "4.4.9"
 ed25519-dalek = "2.1.1"
 codec = { package = "parity-scale-codec", version = "3", default-features = false }
 hex = "0.4.3"
-alloy-primitives = {version = "0.7.4", features = ["serde"]}
+alloy-primitives = { version = "0.7.5", features=["serde"] }
+alloy-sol-types = "0.7.5"
+alloy-sol-macro = "0.7.5"
